api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
websockets==12.0

# parsing
pymupdf==1.24.9
python-docx==1.1.2
beautifulsoup4==4.12.3
trafilatura==1.6.3
Pillow==10.4.0

# embeddings (text + image)
sentence-transformers==3.0.1
torch==2.3.1
torchvision==0.18.1
transformers==4.45.0

# OCR and document preprocessing
accelerate==1.2.1

# vector db
chromadb==0.4.15

# bg tasks
rq==1.16.2
redis==5.0.7

# storage
minio==7.2.7

faster-whisper==1.0.3
numpy==1.26.4
soundfile==0.12.1
